made cds more like cd
You can apparently call "cd -" to go to the last directory that you came from. I wanted to make sure that cds could do the same.
Now, cds encapsulates almost all of the functionality of cd except the use of cd without any parameters, which defaults to the home directory. Instead, cds uses this for displaying a list of the available shortcuts. Maybe I'll fix this later...